restructure trending components
