+++ src/modules/trending/component.jsx
